+++ src/views/Education.vue
@@ -1,14 +1,19 @@
 <template>
   <div>
-    <h1>Hello Education Vue</h1>
-    <p>Vue have {{ stars }} stars on github</p>
+    <Banner class="bar-margin" image-url="/img/banner.svg"></Banner>
+    <div class="container">
+      <h1>Hello Education Vue</h1>
+      <p>Vue have {{ stars }} stars on github</p>
+    </div>
   </div>
 </template>
 
 <script>
 import ApiService from "../services/ApiService";
+import Banner from "../components/Banner";
 
 export default {
+  components: { Banner },
   data() {
     return {
       stars: ""