+++ src/views/Home.vue
@@ -1,5 +1,15 @@
 <template>
   <div>
-    <h1>Hello Vue</h1>
+    <Banner class="bar-margin" image-url="/img/banner.svg"></Banner>
+    <div class="container">
+      <h1>Hello Vue</h1>
+    </div>
   </div>
 </template>
+
+<script>
+import Banner from "../components/Banner";
+export default {
+  components: { Banner }
+};
+</script>