+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### [v0.1.0; 16-08-2019]
 
 #### Added
+- 24/08/2019: New component banner
 - 23/08/2019: New navbar layout
 - 23/08/2019: Update dependencies
 - 16/08/2019: Api Call